Travis only runs the tests for symfit, use pytest
Travis was using nose to run the tests. In addition to the tests for
symfit, nose was also running the tests for sympy, hence the very long
running time.

This commit have Travis run only the tests for symfit. It also switches
from nose to pytest as nose is not maintained anymore.